pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "hammers"
version = "0.0.1"
authors = [{ name = "Chameleon Team", email = "help@chameleoncloud.org" }]
description = 'Bag of hammers to fix problems'
readme = "README.rst"
requires-python = ">=3.9"
classifiers = [
  'Development Status :: 4 - Beta',
  'Environment :: OpenStack',
  'Intended Audience :: System Administrators',
  'Topic :: Utilities',
]

dependencies = [
  "openstacksdk @ git+https://github.com/ChameleonCloud/openstacksdk@chameleoncloud/blazar",
  "oslo.config",
  "oslo.service",
  "oslo.log",
  "iso8601",
  "pyyaml",
]


[project.urls]
Homepage = 'https://github.com/ChameleonCloud/hammers'

[project.optional-dependencies]
dev = [
  "ruff",
  "pytest",
  "freezegun"
  ]

[project.scripts]
periodic_inspector = "hammers.periodic_node_inspector:main"
floating_ip_reaper = "hammers.ip_cleaner:launch_main"
expired_project_reaper = "hammers.expired_project_cleaner:launch_main"
image_deployer = "hammers.image_deployer:launch_main"
set_image_property = "hammers.set_image_property:launch_main"
